requirements.txt
config>=0.5.1
ipykernel>=7.1.0
pandas>=2.3.3
python-dotenv==1.2.1
sqlalchemy==2.0.44